+++ app/controllers/api/v1/base_controller_v1.rb
@@ -1,0 +1,7 @@
+module Api::V1
+  class BaseControllerV1 < ApplicationController
+
+    respond_to :json
+    
+  end
+end