+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-# 1.0.0
+# 0.5.1
+
+* New data on countries endpoints
+
+# 0.5.0
+
+* Add countries endpoints
+
+# 0.5.0
+
+* Setting Up API
+
+# 0.4.0
 
 * Setting Up Bower on Rails for Heroku